Fix problem with standalone
Hi there,
I tried to start the standalone version, but the login form always redirected me to the standalone filename in the manager directory. Obviously the standalone file is not found there...
So I changed line 25, because if you're in standalone mode, the url of the script should be the base url not the manager url.